Emphasize Trello specifics (#15)
IMPORTANT: Review #13 and #14 first.

## Summary

Every pull request consists of:

- Things that are general for every snapin (e.g. "you should do 'data
extraction' phase for ...)
- Things that are 3rd party-service specific

This PR is an aim to make it clear what fits into the first and what
into the second category.

This will make writing the specification more robust.

Every Trello-specific is now prefixed with "Trello Specific: " to make
it clear that when writing specification for new 3rd party service, this
should be removed.

Issue:
- https://app.devrev.ai/devrev/works/ISS-224456

---------

Co-authored-by: Patricija Brečko <patricija.brecko@devrev.ai>